requirements.txt
numpy
pillow
opencv-python
streamlit